+++ test/api/v01/vrp_test.rb
@@ -50,8 +50,8 @@ class Api::V01::VrpTest < Minitest::Test
     OptimizerWrapper.stub(
       :define_main_process,
       lambda { |services_vrps, _job|
-        assert_equal [['skill']], services_vrps[0][:vrp][:vehicles][0][:skills]
-        assert_equal ['skill'], services_vrps[0][:vrp][:services][0][:skills]
+        assert_equal [[:skill]], services_vrps[0][:vrp][:vehicles][0][:skills]
+        assert_equal [:skill], services_vrps[0][:vrp][:services][0][:skills]
         assert_equal [[]], services_vrps[0][:vrp][:vehicles][1][:skills]
         assert_equal [], services_vrps[0][:vrp][:services][1][:skills]
         {}
@@ -244,31 +244,27 @@ class Api::V01::VrpTest < Minitest::Test
   end
 
   def test_block_call_under_clustering
-    @job_ids = []
-    asynchronously start_worker: true do
-      vrp = VRP.lat_lon_scheduling_two_vehicles
-      vrp[:configuration][:preprocessing][:partitions] = TestHelper.vehicle_and_days_partitions
-      @job_ids << submit_vrp(api_key: 'demo', vrp: vrp)
-      response = wait_status @job_ids.last, 'completed', api_key: 'demo'
-      refute_empty response['solutions'].to_a, "Solution is missing from the response body: #{response}"
+    vrp1 = VRP.lat_lon_scheduling_two_vehicles
+    vrp1[:configuration][:preprocessing][:partitions] = TestHelper.vehicle_and_days_partitions
 
-      vrp = VRP.independent_skills
-      vrp[:points] = VRP.lat_lon_scheduling[:points]
-      vrp[:services].first[:skills] = ['D']
-      vrp[:configuration][:preprocessing] = {
-        max_split_size: 4,
-        partitions: [
-          { method: 'balanced_kmeans', metric: 'duration', entity: :vehicle }
-        ]
-      }
-      @job_ids << submit_vrp(api_key: 'demo', vrp: vrp)
-      response = wait_status @job_ids.last, 'completed', api_key: 'demo'
-      refute_empty response['solutions'].to_a, "Solution is missing from the response body: #{response}"
-    end
-
-    @job_ids.each{ |job_id|
-      delete_completed_job job_id, api_key: 'demo'
+    vrp2 = VRP.independent_skills
+    vrp2[:points] = VRP.lat_lon_scheduling[:points]
+    vrp2[:services].first[:skills] = ['D']
+    vrp2[:configuration][:preprocessing] = {
+      max_split_size: 4,
+      partitions: [
+        { method: 'balanced_kmeans', metric: 'duration', entity: :vehicle }
+      ]
     }
+
+    asynchronously start_worker: true do
+      [vrp1, vrp2].each{ |vrp|
+        @job_id = submit_vrp(api_key: 'demo', vrp: vrp)
+        response = wait_status @job_id, 'completed', api_key: 'demo'
+        refute_empty response['solutions'].to_a, "Solution is missing from the response body: #{response}"
+        delete_completed_job @job_id, api_key: 'demo'
+      }
+    end
   end
 
   def test_unfounded_avancement_message_change
@@ -369,5 +365,36 @@ class Api::V01::VrpTest < Minitest::Test
     post '/0.1/vrp/submit', {api_key: 'demo', vrp: vrp_no_sched_start}.to_json, 'CONTENT_TYPE' => 'application/json'
     assert_equal 400, last_response.status
     assert_includes(JSON.parse(last_response.body)['message'], 'vrp[configuration][schedule][range_indices][end] is missing')
+  end
+
+  def test_ask_for_geometry
+    [false, true, # boolean
+     ['polylines'], [:polylines], [:polylines, 'partitions'], ['unexistant'], # array of string or symbol
+     'partitions', 'polylines,partitions', 'polylines, encoded_polylines' # string to be converted into an array
+    ].each_with_index{ |geometry_field, case_index|
+      OptimizerWrapper.stub(:define_main_process, lambda { |services_vrps, _job|
+          case case_index
+          when 0
+            assert_empty services_vrps.first[:vrp].restitution_geometry
+          when 1 || 4 || 7
+            assert_equal %i[polylines partitions], services_vrps.first[:vrp].restitution_geometry
+          when 2 || 3
+            assert_equal %i[polylines], services_vrps.first[:vrp].restitution_geometry
+          when 6
+            assert_equal %i[partitions], services_vrps.first[:vrp].restitution_geometry
+          end
+          {}
+        }
+      ) do
+        vrp = VRP.toy
+        vrp[:configuration][:restitution] = { geometry: geometry_field }
+        if [5, 8].include?(case_index)
+          post '/0.1/vrp/submit', { api_key: 'demo', vrp: vrp }.to_json, 'CONTENT_TYPE' => 'application/json'
+          assert_includes [400], last_response.status
+        else
+          submit_vrp api_key: 'demo', vrp: vrp
+        end
+      end
+    }
   end
 end